Refactoring: Moved DcatSourceDto and UserDto to common module
