+++ libraries/shared/pom.xml
@@ -21,6 +21,12 @@
             <artifactId>jackson-databind</artifactId>
             <version>2.8.6</version>
         </dependency>
+
+      <dependency>
+          <groupId>org.hibernate</groupId>
+          <artifactId>hibernate-validator</artifactId>
+          <version>5.3.5.Final</version>
+      </dependency>
     </dependencies>
 
 </project>
